+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.2
+### Features
+- add scoped slots for customizing keys
+
+# 0.1.1
+### Fix
+- default locale & layout to english(QWERTY)
+
 # 0.1.0
 ### Features
 - Add default actions: toggle typeset